move test files from backend/src/tests to backend/tests
